Day 3 Upload
Organized HTML, CSS, and JS into separate folders. Applied JS to the website under the sandbox html file. Started to work on making a calculator using HTML, CSS, and JS. Spent half of the time learning JS for the first time and the other half applying it to the website.